+++ app/src/main/java/com/fansen/phr/service/IDiagnosisDictService.java
@@ -2,9 +2,11 @@ package com.fansen.phr.service;
 
 import com.fansen.phr.entity.DictDiagnosis;
 
+import java.util.ArrayList;
+
 /**
  * Created by 310078142 on 2015/10/20.
  */
 public interface IDiagnosisDictService {
-    public int addDiagnosisDict(DictDiagnosis dictDiagnosis);
+    public long addDiagnosisDict(DictDiagnosis dictDiagnosis);
 }
